refactor(import): extract supplementary discovery + reporting helpers (epic-4 A2)
Split src/cli/commands/import_data.py (730 lines) under the 500-line
rule by extracting two cohesive modules:

- src/services/firstrate/supplementary_discovery.py: dividend/split
  source-dir auto-discovery and ticker derivation helpers.
- src/cli/commands/import_reporting.py: progress-line and Rich summary
  table rendering helpers.

import_data.py now re-exports the helpers it still calls, so existing
test imports keep working. Result: import_data.py is 456 lines.
